text101/requirements.txt
python-dotenv 
openai
streamlit
google-generativeai
anthropic
cohere 